refactor(oauth): keycardError markers on credential errors; keep deprecated EKS surface in mcp
The moved credential error types now implement the package's
KeycardError marker interface. The deprecated EKS constructor, options,
and error aliases stay in mcp, implemented over oauth's exported
FileTokenSource API, so the oauth package carries no deprecated
surface.
