+++ oauth/credentials.go
@@ -305,63 +305,3 @@ func (w *WebIdentityCredential) ClientID() string {
 func (w *WebIdentityCredential) ClientJWKSURL(resourceServerURL string) string {
 	return w.keyManager.ClientJWKSURL(resourceServerURL)
 }
-
-// EKSWorkloadIdentityCredential authenticates with an AWS EKS pod-identity
-// (projected service-account) token read from a mounted file.
-//
-// Deprecated: use NewWorkloadIdentity with NewFileTokenSource, which also
-// covers AKS and other platforms that project token files.
-type EKSWorkloadIdentityCredential = WorkloadIdentityCredential
-
-// defaultEKSEnvVars is the env-var discovery order for NewEKSWorkloadIdentity.
-// Unlike defaultFileTokenEnvVars it is limited to the EKS variables.
-var defaultEKSEnvVars = []string{
-	"KEYCARD_EKS_WORKLOAD_IDENTITY_TOKEN_FILE",
-	"AWS_CONTAINER_AUTHORIZATION_TOKEN_FILE",
-	"AWS_WEB_IDENTITY_TOKEN_FILE",
-}
-
-// EKSWorkloadIdentityOption configures NewEKSWorkloadIdentity.
-//
-// Deprecated: use FileTokenSourceOption with NewFileTokenSource.
-type EKSWorkloadIdentityOption func(*eksConfig)
-
-type eksConfig struct {
-	tokenFilePath string
-	envVarName    string
-}
-
-// WithTokenFilePath sets the path to the EKS token file directly.
-//
-// Deprecated: use WithFileTokenPath with NewFileTokenSource.
-func WithTokenFilePath(path string) EKSWorkloadIdentityOption {
-	return func(cfg *eksConfig) { cfg.tokenFilePath = path }
-}
-
-// WithEnvVarName adds a custom environment variable name to check for the token file path.
-//
-// Deprecated: use WithFileEnvVar with NewFileTokenSource.
-func WithEnvVarName(name string) EKSWorkloadIdentityOption {
-	return func(cfg *eksConfig) { cfg.envVarName = name }
-}
-
-// NewEKSWorkloadIdentity creates a workload identity credential that reads the
-// EKS projected token file, with env-var discovery limited to the EKS
-// variables.
-//
-// Deprecated: use NewWorkloadIdentity with NewFileTokenSource.
-func NewEKSWorkloadIdentity(opts ...EKSWorkloadIdentityOption) (*EKSWorkloadIdentityCredential, error) {
-	cfg := eksConfig{}
-	for _, opt := range opts {
-		opt(&cfg)
-	}
-
-	source, err := newFileTokenSource(defaultEKSEnvVars, fileTokenSourceConfig{
-		tokenFilePath: cfg.tokenFilePath,
-		envVarName:    cfg.envVarName,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return NewWorkloadIdentity(source)
-}
